+++ api/src/main/java/ar/edu/pod/tp2/Mappers/Query5MapperFirst.java
@@ -6,12 +6,12 @@ import com.hazelcast.mapreduce.Context;
 import com.hazelcast.mapreduce.Mapper;
 
 
-public class Query5MapperFirst implements Mapper<String, SensorReading, String, Integer> {
+public class Query5MapperFirst implements Mapper<String, SensorReading, String, Long> {
 
     @Override
-    public void map(String s, SensorReading sensorReading, Context<String, Integer> context) {
+    public void map(String s, SensorReading sensorReading, Context<String, Long> context) {
         if (sensorReading.getSensorStatus().equals(SensorStatus.A)){
-            context.emit(sensorReading.getSensorDescription(), sensorReading.getHourlyCounts() );
+            context.emit(sensorReading.getSensorDescription(), Long.valueOf(sensorReading.getHourlyCounts()) );
         }
     }
 }
